Modify IProjection::Stitch to take S2Shape::Edge instead of S2Point.
Some projections (such as Equirectangular) need to do orientation checks to
determine which direction a chain crossed a cut in order to connect vertices in
the proper order to maintain interiors.

Generally we need three points to determine orientation around another point,
so provide an entire edge as the first parameter to enable this.